test prune logic

+++ internal/scripts/prune-preview-deploys.ts
@@ -67,7 +67,7 @@ async function ListPreviewWorkerDeployments() {
 
 async function deletePreviewWorkerDeployment(id: string) {
 	const url = `https://api.cloudflare.com/client/v4/accounts/${env.CLOUDFLARE_ACCOUNT_ID}/workers/scripts/${id}`
-	console.log('DELETE', url)
+	nicelog('DELETE', url)
 	const res = await fetch(url, {
 		method: 'DELETE',
 		headers: {
@@ -75,23 +75,47 @@ async function deletePreviewWorkerDeployment(id: string) {
 			'Content-Type': 'application/json',
 		},
 	})
-	console.log('status', res.status)
+	nicelog('status', res.status)
 
 	if (!res.ok) {
 		throw new Error('Failed to delete worker ' + JSON.stringify(await res.json()))
 	}
 }
 
+const neonHeaders = {
+	Authorization: `Bearer ${env.NEON_API_KEY}`,
+}
+
+async function getBranchId(branchName: string) {
+	const url = `https://console.neon.tech/api/v2/projects/${env.NEON_PROJECT_ID}/branches?search=${branchName}`
+	nicelog('GET', url)
+	const res = await fetch(url, {
+		headers: neonHeaders,
+	})
+
+	if (!res.ok) {
+		throw new Error('Failed to list branches ' + JSON.stringify(await res.json()))
+	}
+
+	const data = (await res.json()) as { branches: { id: string; name: string }[] }
+
+	return data.branches.find((b) => b.name === branchName)?.id
+}
+
 async function deletePreviewDatabase(prNumber: number) {
-	const url = `https://console.neon.tech/api/v2/projects/${env.NEON_PROJECT_ID}/branches/pr-${prNumber}`
-	console.log('DELETE', url)
+	const branchName = `pr-${prNumber}`
+	const id = await getBranchId(branchName)
+	if (!id) return
+
+	const url = `https://console.neon.tech/api/v2/projects/${env.NEON_PROJECT_ID}/branches/${id}`
+	nicelog('DELETE', url)
 	const res = await fetch(url, {
 		method: 'DELETE',
 		headers: {
 			Authorization: `Bearer ${env.NEON_API_KEY}`,
 		},
 	})
-	console.log('status', res.status)
+	nicelog('status', res.status)
 }
 
 async function main() {
@@ -111,5 +135,3 @@ async function main() {
 }
 
 main()
-
-// clean up cloudflare preview deploys
